bugs fix

+++ event-notificator/src/main/java/dev/sorokin/eventnotificator/domain/NotificationPayload.java
@@ -8,8 +8,8 @@ public record NotificationPayload(
         String messageId,
         String eventType,
         String occurredAt,
-        Integer changedById,
-        Integer ownerId,
+        Long changedById,
+        Long ownerId,
         String eventName,
         List<ChangeItem> changeItem
 ) {
